Store geonames data as csv.gz, load as csv into bigquery. Moved gzip_file_crc function from test_grid.py to test_utils.py.
